requirements.txt
flask
google-generativeai
python-dotenv
PyPDF2
python-docx
pandas
openpyxl
fpdf2
gunicorn